GUI() get_widget will return an aggregate widget object
Break up WidgetWrapper into a base class and the main class.  GUI.get_widget
will now return a wrapped widget, regardless if the widget has been wrapped.
This provides a consistent interface to acces the widget, the call will 
always have to reference .widget directly.  Hopefully we can eliminate this
with some snazzy metaclass magic - but that is beyond my py-fu as of yet.

Signed-off-by: Darren Hart <darren@dvhart.com>