Instantiate the TaskViewBy object after the tree model
TaskViewBy makes use of the view_by* methods of the tree model, so
wait until the model exists before creating it.  Perhaps a better approach
would be to prevent TaskViewBy from using the methods in the constructor,
or perhaps requiring an instantiated model object as an argument to the
constructor.  But this fix removes the errors for the time being.

Signed-off-by: Darren Hart <darren@dvhart.com>