Add the ContextTable widget
The widget provides basic functionality for setting task contexts.  It
automatically arranges the context checkbuttons as it is resized - or it
will once I can figure our configure_events.  It is current bound to the
configure_events of the main window, that will be removed once they are
working for the eventbox in which the table is placed.

Signed-off-by: Darren Hart <darren@dvhart.com>