Remove premature datastructure optimization
I had tried to minimize some looping and task aggregation by storing
tasks in contexts and in projects.  As I tried to flesh out the app
logic I'm realizing that I should leave the structure simple for now
and optimize later, after the logic is mostly completed and I can see
where the most cycles are being spent.  This simplifies the actual
structure, and adds a sudo-recursive get_tasks() call to the hierarchy.
It is certainly less efficient, but I think it will help move toward
a stable gtd Tree API and aid in moving the app logic along.

Signed-off-by: Darren Hart <darren@dvhart.com>