ContextTable auto-resize
Turns out configure_events are only emitted for windows and
drawing_areas, so the eventbox won't work there.  I was able to
reconfigure the table layout with the size_allocate() signal by
adding a last_allocation flag to avoid the 
size_allocate()->attach()->size_allocate()... loop.  We basically
negotiate a size with GTK and once it stays the same (or at least
the rows, cols do) we end the cycle.  This is a bit jerkier than I
would like.  Gimp uses a custom wbox (wrap box) to create the toolbox
widget.  It is a completely new container and is quite involved, while
this modification is much simpler.  I think is adequate for the time being.

Signed-off-by: Darren Hart <darren@dvhart.com>