Update the tree when the context table is updated
This patch adds the methods to update the tree and model when the
contexts in the context table are toggled.  It also stubs in a
similar routine for when the project combo box changes.  This is
mostly functional, other I think some object interaction review
is in order.  I am not happy with the TreeView and TreeModel setup.
At the moment the rest of the app interfaces with the TreeView, which
may be the right way to go... but we'll see... :-)

Signed-off-by: Darren Hart <darren@dvhart.com>