Major layout overhaul - separate filter and task lists
This patch provides a significant funtional improvement of the UI.
It reduces duplicate entries by not displaying a context/task tree, and
facilitates work by allowing the user to filter by easily selectable
projects and contexts.  Some connections still need to be made (or moved)
and the project as a whole would benefit from a signals and slots system
on the gtd.Tree object itself.  Then each object can register it's own
interest in the underlying data, rather than making the other objects aware
of eachother.  Seems like a much more robust and maintainable model to me.

Signed-off-by: Darren Hart <darren@dvhart.com>