Remove the task_title edit box in favor of a special row in the task_list
Make use of the NewTask class to list a special row in the task_list
that creates (or will...) a new task when it's title is changed.  The new
tags filter code uses a simple list (rather than a tree) so we only need
one such row, rather than one per tag.  So this approach now simplifies the
gui by reducing redundancy, and is also self documenting.  It also removes
the need for an "Update" "Add" set of buttons and the confusion those can
bring in.  I relly like this change!

Signed-off-by: Darren Hart <darren@dvhart.com>