Tie in the GTDNone type objects.
These objects are here to allow the creation of tasks without a complete
hierarchy, as people are apt to do.  This full path allows us to consolidate
the behavior for the no hierarchy case in these classes, rather than checking for
a None parent every place we access the parent.

Signed-off-by: Darren Hart <darren@dvhart.com>