Initial use of Signals and Slots from PyNotify
This introduces a dependency on PyNotify, but the library seems to be quite
complete and provide a much needed mechanism.  The TreeListener interface
has removed in favor of these far more flexible signals.

The current implementation makes use of the new OProperty code, which allows
for inheritance with properties.  We could avoid that by having the Base
object call an on_rename() signal which is set to the appropriate signal
in __init__ of each derived class.  I think OProperty is generally useful
though, so am taking this approach for now.

So far only added tree-level signals: renamed, added, removed for each gtd
type, plus visible_changed for the realms.  It may make sense to push some of
these signals down to the individual gtd types, but I think we'll start here
and see how much that may be needed.  This implementation will certainly cause
more O(N) searches, but given the amount of data we'll be dealing with for the
most part, I'm not sure that's even an issue.

Patch includes a new GTDSignalTest class to test the delivery of the signals.

Signed-off-by: Darren Hart <darren@dvhart.com>