Use application-wide datastores
Move to using single datastores for each gtd type.  Then create filters from those for
use in the various widgets as necessary.  This really simplifies the datamodel, we only
have to let the primary datastores know when something changed, and they do the rest, refiltering
all the "child" filter datastores.

TaskStore is still yet to be implemented.

Signed-off-by: Darren Hart <darren@dvhart.com>