Realm Area Dialog
This is the start of the Realms and Areas Dialog.  It will replace the
RealmToolEntry widget as the means to add and rename Realms.  This will
also be used to add areas, and makes it clear which realm a new area
will belong to.  Removing realm and area management from the main window
and placing it a dialog makes sense as it is not a common task.

o The RealmAreaStore still needs a lot of work.
o The drag and drop behavior in the RealmAreaTreeView needs to be thought 
  through.

Signed-off-by: Darren Hart <darren@dvhart.com>