Add new contexts and realms to the top level GTD Tree
This patch adds two new internal methods to the gtd class for adding
new contexts and realms to the tree.  Theses are top level elements
and connect directly to the class GTD, rather than another base GTD
element, so they require some special casing.  Since we're calling
into the GTD class anyway, I thought it made sense to emit the signals
from there as well... but this differes from the the __init__s of the
other gtd elements, so might lead to confusion... I'll give that some 
thought.

Signed-off-by: Darren Hart <darren@dvhart.com>