Initial Infrastructure for adding dates to the application
This patch adds start and due dates to tasks and projects.  It adds a 
combo box to filter by all,active,future, and someday tasks/projects
from the task/project tabs.  The tie in to the listview model filters
is a real ugly hack.  Some better integration with the selection
filters is needed.

Signed-off-by: Darren Hart <dvhltc@us.ibm.com>