Resize the window when the panels are hidden
This is probably usually the right thing since users are likely to keep the
UI small.  However, if the window is not forced to get bigger to display
the panels in the first place, we probably shouldn't shrink when they are
destroyed.  (other apps don't resize when you close their sidebar...)

Signed-off-by: Darren Hart <darren@dvhart.com>