Add an optional due_date sort to GTDStore
Unfortunately, the sort function is set on the model, not the view.  This
makes it necessary to duplicate the model to change how things are sorted.
The project filter view should be sorted alphabetically.  The project working
list (main window) should be sorted by due_date, then by title, requiring a
second ProjectStore().  Grumble.

Signed-off-by: Darren Hart <darren@dvhart.com>