fix new task's being created with "No Context"
The new task handler in braindump.py would blindly pass a ContextNone in the
contexts array to the Task constructor.  This would get saved in the xmlstore,
and when subsequently loaded, the reference would never get resolved as a new
ContextNone is created with each application start (and is not saved in the
xmlstore).  We represent ContextNone in the Task objects with an empty
contexts array.  I'd like to scrap all the gtd.*None classes eventually.

Signed-off-by: Darren Hart <darren@dvhart.com>