Use the model for the countdown_frame
Now that we have access to the model and iter in the filters, use the model to
store the countdown_frame.  Since this is a purely UI element, it makes more
sense to store it here, rather than in the gtd domain objects.

Add a GTDStore.add() method to avoid using self.model.append([obj, 0]) outside
of GTDStore. (The 0 being the countdown frame).

Signed-off-by: Darren Hart <darren@dvhart.com>