Restore icon clear with gtk 2.16 secondary icon and icon-release signal
Update the glade file to use the gtk-clear stock icon in the secondary icon
location. Map the icon-release signal to on_search_clear.

Restructure SearchEntry to work with self.widget as the actual gtk.Entry instead
of a gtk.Alignment container widget.

Rework the logic as it behaved eratically before. The whole _active and _focused
concept should be thrown out and redesigned. It's ugly. But that's another
patch.

Update the README to reflect a dependency on gtk 2.16.

Signed-off-by: Darren Hart <darren@dvhart.com>